fix autoflake issues.
